requirements.txt
asgiref==3.7.2
Django==5.0.1
django-jasmine==0.5.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
